Resolve all lint warnings and refactor component structure
- Fixed all remaining lint warnings (8 → 0)
  - Resolved non-null assertion warnings with proper null checks
  - Fixed React Fast Refresh warnings by separating components
  - Removed unused imports

- Refactored expressions.tsx and statements.tsx into modular structure
  - Split into subdirectories with separate files for each component
  - Used factory pattern to avoid circular dependencies
  - Separated React components from dispatcher functions
  - Properly typed all components and helper functions

- Improved file organization
  - expressions/ and statements/ subdirectories
  - Separate files for types, components, dispatchers
  - Used .tsx for files with JSX, .ts for pure TypeScript

All lint checks pass with 0 errors and 0 warnings.
TypeScript compilation successful with no errors.
