+++ src/views/expressions/Identifier.tsx
@@ -1,0 +1,10 @@
+import React, { ReactNode } from 'react';
+import { ExpressionProps } from './types';
+
+export const Identifier: React.FC<ExpressionProps> = ({ node }) => {
+  return (
+    <span className="expression identifier">
+      {node.get('name') as ReactNode}
+    </span>
+  );
+};
