+++ src/views/expressions/ThisExpression.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { ExpressionProps } from './types';
+
+export const ThisExpression: React.FC<ExpressionProps> = () => {
+  return (
+    <span className="expression this-expression">
+      <span className="keyword">this</span>
+    </span>
+  );
+};
