+++ src/views/expressions/index.ts
@@ -1,0 +1,5 @@
+// Export only the dispatcher functions
+export { dispatchExpression, dispatchPattern } from './dispatchers';
+
+// Export types if needed
+export type { ExpressionProps, ExpressionComponent } from './components';
