+++ src/views/statements/index.ts
@@ -1,0 +1,5 @@
+// Export the dispatcher and statement components
+export { dispatchStatement, statementComponents } from './dispatchers';
+
+// Export types if needed
+export type { StatementProps, StatementComponent } from './types';
